Updated Wire.ByEdges
Wire.ByEdges now does not require a head to tail edges. It can create any wire (open or closed) from a set of intersecting edges. The only requirement is that the edges (in total) do intersect to create one connected wire. If the edges are not intersecting, you should use Cluster.ByTopologies instead.