Updated GlobalCluster.Clear
GlobalCluster.Clear now takes a an input named "Wait For" which forces the clear operation to wait for other nodes to finish. Just connect to it any outputs that you need to have executed before the clear node tirggers. The node does not process the "Wait For" input in any way.